向日葵

+++ src/main/java/com/hungteen/pvz/utils/AnimationUtil.java
@@ -7,39 +7,39 @@ public class AnimationUtil {
 	private static final float PI = 3.1415926535F;
 	private static final float ANGLE_TO = PI / 180;
 	
-	public static float upDown(int x, int t, float scale) {
+	public static float upDown(float x, float t, float scale) {
 		final float sita = PI / t;
 		return MathHelper.abs(MathHelper.sin(sita * x)) * scale;
 	}
 	
-	public static float upDownUpDown(int x, int t, float scale) {
+	public static float upDownUpDown(float x, float t, float scale) {
 		final float sita = 2 * PI / t;
 		return MathHelper.sin(sita * x) * scale;
 	}
 	
-	public static float up(int x, int t, float scale) {
+	public static float up(float x, float t, float scale) {
 		final float sita = PI / 2 / t;
 		return MathHelper.sin(sita * x) * scale;
 	}
 	
-	public static float down(int x, int t, float scale) {
+	public static float down(float x, float t, float scale) {
 		final float sita = PI / 2 / t;
 		return MathHelper.cos(sita * x) * scale;
 	}
 	
-	public static float getUpDown(int x, int t, float maxAngle) {
+	public static float getUpDown(float x, float t, float maxAngle) {
 		return upDown(x, t, maxAngle * ANGLE_TO);
 	}
 	
-	public static float getUpDownUpDown(int x, int t, float maxAngle) {
+	public static float getUpDownUpDown(float x, float t, float maxAngle) {
 		return upDownUpDown(x, t, maxAngle * ANGLE_TO);
 	}
 	
-	public static float getUp(int x, int t, float maxAngle) {
+	public static float getUp(float x, float t, float maxAngle) {
 		return up(x, t, maxAngle * ANGLE_TO);
 	}
 	
-	public static float getDown(int x, int t, float maxAngle) {
+	public static float getDown(float x, float t, float maxAngle) {
 		return down(x, t, maxAngle * ANGLE_TO);
 	}
